refactor: split diplomacy phase-result value types per file (Refs #3419)
Step 9 of the diplomacy dedup/refactor: move the per-interaction human-input
value types out of diplomacy_phase_result.dart into one file per interaction
family under src/diplomacy/phase_types/ (overture, ftp, intervention,
call-to-arms). diplomacy_phase_result.dart retains DiplomacyPhaseResult and
re-exports the new files, preserving the package barrel surface and the turn
package's narrow re-export of this file (turn -> diplomacy one-way edge).

Behaviour-preserving. Adds diplomacy_phase_types_split_test.dart gating that
each new file is self-contained (deep-import construction) and the aggregator
re-export resolves to the same types.
